requirements.txt
python-telegram-bot==21.4
caldav==1.3.9
icalendar==5.0.13
python-dotenv==1.0.1
